modules version of code
