+++ index.html
@@ -25,7 +25,6 @@
     <a href="index.html">Reset</a>
     <a href="pages/rankings.html" target="_blank">Rankings</a>
     <a href="pages/about.html" target="_blank">About</a>
-    <span id="team-name-display" class="team-name"></span>
   </nav>
   <div class="main-container">
     <div class="container" id="login-container">
@@ -39,39 +38,7 @@
 
       <p id="login-error" class="hidden"></p>
     </div>
-    <div class="container hidden" id="quiz-container">
-      <h1>Science and Tech Trivia</h1>
-      <p id="score-display">Score: 0</p>
-      <p id="countdown-display">Time left: 10</p>
-      <div class="question-container">
-        <p id="question"></p>
-        <img id="question-image" src="" alt="Question related image" style="width: 100%; display: none;">
-        <div id="multiple_choice_options">
-          <div>
-            <input type="radio" name="option" value="a" id="option_a">
-            <label for="option_a" id="option-a"></label>
-          </div>
-          <div>
-            <input type="radio" name="option" value="b" id="option_b">
-            <label for="option_b" id="option-b"></label>
-          </div>
-          <div>
-            <input type="radio" name="option" value="c" id="option_c">
-            <label for="option_c" id="option-c"></label>
-          </div>
-          <div>
-            <input type="radio" name="option" value="d" id="option_d">
-            <label for="option_d" id="option-d"></label>
-          </div>
-        </div>
-        <div id="short_answer_input" class="hidden">
-          <input type="text" placeholder="Your answer" id="short_answer">
-        </div>
-        <button onclick="submitAnswer()">Submit</button>
-      </div>
-      <p id="result" class="hidden"></p>
-    </div>
   </div>
-  <script src="js/main.js"></script>
+  <script src="js/login.js"></script>
 </body>
 </html>
