+++ pages/quiz.html
@@ -1,0 +1,71 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Science and Tech Trivia</title>
+  <link rel="stylesheet" href="../css/main.css">
+  <script>
+    const setBaseURL = () => {
+      let baseurl;
+      if (window.location.hostname === "sanjin84.github.io") {
+        baseurl = "/AITrivia";
+      } else {
+        baseurl = "";
+      }
+      const base = document.createElement("base");
+      base.href = baseurl + "/";
+      document.head.appendChild(base);
+    };
+    setBaseURL();
+  </script>
+</head>
+
+<body>
+  <nav class="navbar">
+    <a href="index.html">Reset</a>
+    <a href="pages/rankings.html" target="_blank">Rankings</a>
+    <a href="pages/about.html" target="_blank">About</a>
+    <span id="team-name-display" class="team-name"></span>
+  </nav>
+  <div class="main-container">
+    <div class="container" id="quiz-container">
+      <h1>Science and Tech Trivia</h1>
+      <p id="score-display">Score: 0</p>
+      <p id="countdown-display">Time left: 10</p>
+      <div class="question-container">
+        <p id="question"></p>
+        <img id="question-image" src="" alt="Question related image" style="width: 100%; display: none;">
+        <div id="multiple_choice_options">
+          <div>
+            <input type="radio" name="option" value="a" id="option_a">
+            <label for="option_a" id="option-a"></label>
+          </div>
+          <div>
+            <input type="radio" name="option" value="b" id="option_b">
+            <label for="option_b" id="option-b"></label>
+          </div>
+          <div>
+            <input type="radio" name="option" value="c" id="option_c">
+            <label for="option_c" id="option-c"></label>
+          </div>
+          <div>
+            <input type="radio" name="option" value="d" id="option_d">
+            <label for="option_d" id="option-d"></label>
+          </div>
+        </div>
+        <div id="short_answer_input" class="hidden">
+          <input type="text" placeholder="Your answer" id="short_answer">
+        </div>
+        <button id="submit-button">Submit</button>
+      </div>
+        <div id="result-display" class="hidden">
+          <h1 id="result-message"></h1>
+          <p id="correct-answer"></p>
+        </div>
+      <p id="result" class="hidden"></p>
+    </div>
+  </div>
+  <script type="module" src="../js/quiz-controller.js"></script>
+</body>
+</html>
