+++ pages/rankings.html
@@ -18,7 +18,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Ranking Table</title>
-  <link rel="stylesheet" href="css/rankings.css">
+  <link rel="stylesheet" href="../css/rankings.css">
 </head>
 <body>
   <nav class="navbar">
@@ -31,6 +31,6 @@
   <section class="container main-container">
     <table id="ranking-table" class="ranking-table"></table>
   </section>
-  <script src="js/rankings.js"></script>
+  <script src="../js/rankings.js"></script>
 </body>
 </html>
